+++ config/routes.rb
@@ -29,4 +29,6 @@ Rails.application.routes.draw do
   get "up", to: "rails/health#show", as: :rails_health_check
 
   root "home#index"
+
+  draw :application if Rails.root.join("application/config/routes/application.rb").file?
 end